pyproject.toml
[project]
name = "queryleaf"
version = "0.1.0"
description = "Book Companion System"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-openai>=0.3.17",
    "llama-index>=0.12.36",
    "pymupdf>=1.25.5",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.2",
]
